+++ gui-builder/src/GuiBuilder.cpp
@@ -359,7 +359,6 @@ void GuiBuilder::mainLoop()
             else if (event.type == sf::Event::Resized)
             {
                 m_window.setView(sf::View({0.f, 0.f, static_cast<float>(event.size.width), static_cast<float>(event.size.height)}));
-                m_gui.setView(m_window.getView());
             }
             else if (event.type == sf::Event::MouseButtonReleased)
             {
@@ -773,7 +772,6 @@ void GuiBuilder::closeForm(Form* form)
             else if (event.type == sf::Event::Resized)
             {
                 m_window.setView(sf::View({0.f, 0.f, static_cast<float>(event.size.width), static_cast<float>(event.size.height)}));
-                m_gui.setView(m_window.getView());
             }
 
             m_gui.handleEvent(event);