+++ changelog.md
@@ -9,6 +9,7 @@ TGUI 0.9-dev
 - Replaced all std::string and sf::String by tgui::String
 - Replaced Text, Color, Rect and Vector2 classes from SFML with own versions
 - Replaced sf::Text::Style with tgui::TextStyle
+- Replaced setView in Gui by view/viewport setters with absolute or relative values
 - Signal names, renderer names and widget/renderer properties are now case-sensitive
 - Added setWidth and setHeight helper functions
 - Added getCheckedRadioButton to RadioButtonGroup